backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "fastapi-crons>=2.2.0",
    "firebase-admin>=7.3.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "redis>=4.6.0",
    "uvicorn>=0.43.0",
]
